+++ src/entities/club/api/clubQueries.ts
@@ -7,9 +7,6 @@ import {
 } from "@tanstack/react-query";
 import { getClubs } from "./getClubs";
 import { getClubDetail } from "./getClub";
-import { getClubOpeningRequests } from "./getClubOpeningRequests";
-import { getClubOpeningStatus } from "./getClubOpeningStatus";
-import { patchClubApproval } from "./patchClubApproval";
 import { postClub } from "./postClub";
 import { deleteClub } from "./deleteClub";
 import { putClub } from "./putClub";
@@ -27,36 +24,7 @@ export const clubQueries = {
       queryKey: ["club", "detail", id],
       queryFn: () => getClubDetail(id),
     }),
-
-  openingRequests: () =>
-    queryOptions({
-      queryKey: ["club", "opening-requests"],
-      queryFn: getClubOpeningRequests,
-    }),
-
-  openingStatus: () =>
-    queryOptions({
-      queryKey: ["club", "opening-status"],
-      queryFn: getClubOpeningStatus,
-    }),
 } as const;
-
-export const usePatchClubApproval = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({
-      clubId,
-      body,
-    }: {
-      clubId: number;
-      body: { approved: boolean };
-    }) => patchClubApproval(clubId, body),
-    onSuccess: (_, { clubId }) => {
-      queryClient.invalidateQueries({ queryKey: ["club"] });
-      queryClient.invalidateQueries({ queryKey: ["club", "detail", clubId] });
-    },
-  });
-};
 
 export const useDeleteClub = () => {
   const queryClient = useQueryClient();
